- Added configurable burnout - More stable in general - Performance improvements - Can now work with no additional workers - Increased default burnout, might take more iterations for convergence, but faster ones.

+++ src/distributions/multinomial_dist.jl
@@ -2,7 +2,7 @@ using LinearAlgebra
 using Distributions
 
 struct multinomial_dist <: distibution_sample
-    α::AbstractArray{Float64,1}
+    α::AbstractArray{Float32,1}
 end
